+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 
 - Corrected a scheduling error where the `generate` input with a descriptor interval (`@hourly`, etc) had a chance of firing twice.
 - Fixed an issue where a `redis_streams` input that is rejected from read attempts enters a reconnect loop without backoff.
+- The `sqs` input now periodically refreshes the visibility timeout of messages that take a significant amount of time to process.
 
 ### Changed